requirements.txt
jsonpickle==2.2.0
python-dateutil==2.8.2
pytz==2022.6
google==3.0.0
flask==2.2.2
requests==2.28.1
geopy==2.3.0
pycountry==22.3.5
rich==12.6.0
pydantic==1.10.2
prettytable==3.5.0